able to move div elements in the classroom, next need to record their x and y coordinates, save it to backend. Then test performance
