+++ client/src/App.jsx
@@ -3,12 +3,6 @@ import { Routes, Route } from "react-router-dom";
 import TeacherHome from "./pages/TeacherHome";
 import StudentHome from "./pages/StudentHome";
 import Home from "./pages/Home";
-import SubEmotionAnxious from "./pages/student/SubEmotionAnxious"
-import SubEmotionAngry from "./pages/student/SubEmotionAngry"
-import SubEmotionHappy from "./pages/student/SubEmotionHappy";
-import SubEmotionProud from "./pages/student/SubEmotionProud";
-import SubEmotionSad from "./pages/student/SubEmotionSad";
-import SubEmotionScared from "./pages/student/SubEmotionScared";
 import RegZone from "./pages/student/RegZone"
 import GoalsNeeds from "./pages/student/GoalsNeeds"
 import Summary from "./pages/student/Summary"
@@ -22,6 +16,7 @@ import { UserProvider } from "./context/UserContext";
 import ViewClassroom from "./pages/teacher/ViewClassroom";
 import SubEmotion from "./pages/student/SubEmotion";
 import EditTeacher from "./pages/teacher/EditTeacher";
+import TESTEditSeatingChart from "./pages/teacher/TESTEditSeatingChart";
 
 
 export default function App() {
@@ -39,25 +34,20 @@ export default function App() {
             {/* Teacher Routes */}
             <Route path="/teacher-home" element={<TeacherHome />} />
             <Route path="/classroom/:teacherId/:classroomId" element={<ViewClassroom />} />
+            <Route path="/TESTEditSC/:teacherId/:classroomId" element={<TESTEditSeatingChart />} />
             <Route path="/viewclasslist/:teacherId/:classroomId" element={<ClassList />} />
             <Route path="/editneedsgoals" element={<NeedsGoals />} />
             <Route path="/:teacherId/:classroomId/:studentId" element={<StudentProfile />} />
             <Route path="/edit/:teacherId" element={<EditTeacher />} />
 
-        {/* Student Routes */}
-        <Route path="/student-home" element={<StudentHome />} />
-        <Route path="/emotion" element={<SubEmotion/>} />
-        <Route path="/subemotionnervous" element={<SubEmotionAnxious />} />
-        <Route path="/subemotionangry" element={<SubEmotionAngry />} />
-        <Route path="/subemotionhappy" element={<SubEmotionHappy />} />
-        <Route path="/subemotionproud" element={<SubEmotionProud />} />
-        <Route path="/subemotionsad" element={<SubEmotionSad />} />
-        <Route path="/subemotionscared" element={<SubEmotionScared />} />
-        <Route path="/regzone" element={<RegZone />} />
-        <Route path="/goalsneeds" element={<GoalsNeeds />} />
-        <Route path="/summary" element={<Summary />} />
-      </Routes>
-      </UserProvider>
+            {/* Student Routes */}
+            <Route path="/student-home" element={<StudentHome />} />
+            <Route path="/emotion" element={<SubEmotion/>} />
+            <Route path="/regzone" element={<RegZone />} />
+            <Route path="/goalsneeds" element={<GoalsNeeds />} />
+            <Route path="/summary" element={<Summary />} />
+          </Routes>
+        </UserProvider>
       </AuthProvider>
 
     </main>
