+++ client/src/data/emotionData.js
